refactor: stat_value is now double

+++ backend/src/main/java/edu/ntnu/idatt2106_2023_06/backend/dto/stat/StatAddDTO.java
@@ -8,7 +8,7 @@ import lombok.NonNull;
 public record StatAddDTO(@NonNull Long fridgeId,
                          @NonNull Long statTypeId,
                          @NonNull Long userId,
-                         @NonNull Integer statValue,
+                         @NonNull Double statValue,
                          @NonNull Integer quantity,
                          @NonNull String itemName,
                          @NonNull String storeName) {
