+++ backend/src/main/java/edu/ntnu/idatt2106_2023_06/backend/dto/stat/StatDTO.java
@@ -5,7 +5,7 @@ import lombok.NonNull;
 import java.util.Date;
 
 public record StatDTO(@NonNull Long statId,
-                      @NonNull Integer statValue,
+                      @NonNull Double statValue,
                       @NonNull Date timestamp,
                       @NonNull Long fridgeId,
                       @NonNull Long statTypeId,
