+++ backend/src/main/java/edu/ntnu/idatt2106_2023_06/backend/dto/stat/StatDeleteFromFridgeDTO.java
@@ -1,7 +1,7 @@
 package edu.ntnu.idatt2106_2023_06.backend.dto.stat;
 
 public record StatDeleteFromFridgeDTO(Integer percentageThrown,
-                                      Integer price,
+                                      Double price,
                                       Integer quantity,
                                       String itemName,
                                       String storeName,
